Organize source code in separate files and remove unnecessary function
No need to issue a separate request to process repository HTML page.
Instead, since every repository page generally has a issues HTML page,
directly append the '\issues' to the repository page hyperlink and
request that and parse it. So, remove the function 'process_repo_page'.

Some repositories might not have issues page. In that case request to
issues page will still work and a pull requests HTML page is returned.
So, the code ends up saving the hyperlinks of PRs to json file.